Rename and move integration test files
